Add files via upload
"As the project is in the final stage, management has asked to automate the integration and deployment of the web application. One is  required to set up an environment where the application will be hosted and accessed by users. The source code is supposed to be fetched from a GitHub repository."